fix: исправление багов
